+++ src/lib/ultra/gu/coss.c
@@ -1,6 +1,6 @@
 #include "guint.h"
 
-signed short coss(unsigned short x)
+signed int coss(unsigned int x)
 {
-	return sins((unsigned short) (x + 0x4000));
+	return sins((unsigned int) (x + 0x4000));
 }
